// packages/lego-webapp/redux/models/Event.ts
export enum EventType {
  ALTERNATIVE_PRESENTATION = 'alternative_presentation',
  BREAKFAST_TALK = 'breakfast_talk',
  COMPANY_PRESENTATION = 'company_presentation',
  COURSE = 'course',
  EVENT = 'event',
  NEXUS_EVENT = 'nexus_event',
  LUNCH_PRESENTATION = 'lunch_presentation',
  OTHER = 'other',
  PARTY = 'party',
  SOCIAL = 'social',
}

export enum EventStatusType {
  NORMAL = 'NORMAL',
  OPEN = 'OPEN',
  TBA = 'TBA',
  INFINITE = 'INFINITE',
}

export interface CompleteEvent {
  id: EntityId;
  title: string;
  slug: string;
  description: string;
  cover: string;
  coverPlaceholder: string;
  text: string;
  eventType: EventType;
  eventStatusType: EventStatusType;
  location: string;
  comments: EntityId[];
  contentTarget: ContentTarget;
  startTime: Dateish;
  endTime: Dateish;
  mergeTime?: Dateish;
  thumbnail: string;
  pools: EntityId[];
  totalCapacity?: number;
  registrationCloseTime?: Dateish;
  registrationDeadlineHours?: number;
  unregistrationCloseTime?: Dateish;
  unregistrationDeadline?: Dateish;
  unregistrationDeadlineHours?: number;
  company?: ListCompany;
  responsibleGroup?: PublicGroup;
  activeCapacity?: number;
  feedbackDescription: string;
  feedbackRequired: boolean;
  isPriced: boolean;
  priceMember: number;
  priceGuest?: number;
  useStripe: boolean;
  paymentDueDate?: Dateish;
  useCaptcha: boolean;
  waitingRegistrations?: EntityId[];
  waitingRegistrationCount?: number;
  tags: string[];
  isMerged: boolean;
  heedPenalties: boolean;
  createdBy?: PublicUser;
  registrationCount?: number;
  legacyRegistrationCount?: number;
  survey?: EntityId;
  useConsent: boolean;
  youtubeUrl: string;
  mazemapPoi?: number;
  pinned: boolean;
  responsibleUsers: PublicUser[];
  isForeignLanguage: boolean;
  unregistered: EntityId[];
  userReg?: ReadRegistration;

  // for survey
  attendedCount: number;

  // user specific
  price: number;
  activationTime: Dateish;
  isAdmitted: boolean;
  following: false | EntityId;
  spotsLeft: number;
  pendingRegistration?: EntityId;
  photoConsents: PhotoConsent[];

  unansweredSurveys: EntityId[];

  actionGrant: ActionGrant;
}

export type PublicEvent = Pick<
  CompleteEvent,
  | 'id'
  | 'slug'
  | 'title'
  | 'description'
  | 'eventType'
  | 'eventStatusType'
  | 'location'
  | 'thumbnail'
  | 'actionGrant'
>;

// This corresponds to EventReadSerializer in the backend
export type ListEvent = Pick<
  CompleteEvent,
  | 'id'
  | 'title'
  | 'slug'
  | 'description'
  | 'cover'
  | 'coverPlaceholder'
  | 'eventType'
  | 'eventStatusType'
  | 'location'
  | 'startTime'
  | 'endTime'
  | 'thumbnail'
  | 'totalCapacity'
  | 'company'
  | 'registrationCount'
  | 'tags'
  | 'activationTime'
  | 'isAdmitted'
  | 'survey'
  | 'responsibleUsers'
  | 'actionGrant'
  | 'userReg'
> &
  ObjectPermissionsMixin;
// ...
export type EventForSurvey = Pick<
  CompleteEvent,
  'registrationCount' | 'waitingRegistrationCount' | 'attendedCount'
> &
  ListEvent;